Project_37/requirements.txt
markdown-it-py==4.0.0
mdurl==0.1.2
pillow==12.1.0
Pygments==2.19.2
rich==14.2.0
